finito

+++ basegraphics.cpp
@@ -13,9 +13,9 @@
 #include "card.hpp"
 #include "hand.hpp"
 
-namespace el{
+namespace el {
 void DrawText(sf::RenderWindow& window, sf::Font& font, const std::string& str,
-              float x, float y, int size, sf::Color color,
+              float x, float y, unsigned int size, sf::Color color,
               float angle_of_rotation) {
   sf::Text text(str, font, size);
   text.setFillColor(color);
@@ -33,9 +33,9 @@ void drawCircle(sf::RenderWindow& window, float x, float y, float radius,
   circle.setOutlineColor(borders);
   window.draw(circle);
 }
-sf::RectangleShape RectangularButton(float x, float y,
-                                     float w, float h, sf::Color background,
-                                     float thickness, sf::Color borders,
+sf::RectangleShape RectangularButton(float x, float y, float w, float h,
+                                     sf::Color background, float thickness,
+                                     sf::Color borders,
                                      float angle_of_rotation) {
   sf::RectangleShape rect(sf::Vector2f(w, h));
   rect.setPosition(x, y);
@@ -68,7 +68,7 @@ std::vector<sf::Text> createCurvedText(const std::string& text,
   float angleStep = totalAngleDeg / static_cast<float>(text.size() - 1);
 
   for (std::size_t i = 0; i < text.size(); ++i) {
-    float angleDeg = startAngleDeg - i * angleStep;
+    float angleDeg = startAngleDeg - static_cast<float>(i) * angleStep;
     float angleRad = angleDeg * 3.14159265f / 180.f;
 
     float x = center.x + radius * std::cos(angleRad);
@@ -87,4 +87,4 @@ std::vector<sf::Text> createCurvedText(const std::string& text,
 
   return letters;
 }
-}
+}
